refactor instructor group components into their own namespace.
long overdue. will also make it easier to add/switch out sub-components.
