+++ app/components/instructor-group/header.hbs
@@ -1,6 +1,6 @@
 <div
-  class="instructorgroup-header"
-  data-test-instructorgroup-header
+  class="instructor-group-header"
+  data-test-instructor-group-header
   {{did-insert this.load}}
   {{did-update this.load @instructorGroup.title}}
   ...attributes
