+++ app/components/instructor-group/root.hbs
@@ -1,9 +1,9 @@
 <div
-  class="instructorgroup-details"
-  data-test-instructorgroup-details
+  class="instructor-group-root"
+  data-test-instructor-group-root
   ...attributes
 >
-  <InstructorgroupHeader
+  <InstructorGroup::Header
     @instructorGroup={{@instructorGroup}}
     @canUpdate={{@canUpdate}}
   />
