+++ app/styles/components/instructor-group/header.scss
@@ -1,4 +1,4 @@
-.instructorgroup-header {
+.instructor-group-header {
   .header-bar {
     border-bottom: 1px solid $culturedGrey;
     display: flex;
