+++ app/styles/components/instructor-group/root.scss
@@ -1,4 +1,4 @@
-.instructorgroup-details {
+.instructor-group-root {
   padding: .5rem .25rem;
 
   .backtolink {
