+++ app/templates/instructor-group.hbs
@@ -1,2 +1,2 @@
 {{page-title (t "general.instructorGroups")}}
-<InstructorgroupDetails @instructorGroup={{@model}} @canUpdate={{this.canUpdate}} />
+<InstructorGroup::Root @instructorGroup={{@model}} @canUpdate={{this.canUpdate}} />
