+++ tests/pages/instructor-group.js
@@ -1,5 +1,5 @@
 import { create, visitable } from 'ember-cli-page-object';
-import details from './components/instructorgroup-details';
+import details from './components/instructor-group/root';
 
 export default create({
   visit: visitable('/instructorgroups/:instructorGroupId'),
